app/requirements.txt
flask==3.1.0
playwright==1.51.0
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
openpyxl==3.1.2
